dt-bindings: Add binding for the Analogix ANX7688 chip
The ANX7688 chip is a Type-C Port Controller, HDMI to DP converter and
USB-C mux between USB 3.0 lanes and the DP output.

For our use case a big part of the chip, like power supplies, control
gpios and the usb-c part is managed by an Embedded Controller, hence,
this is its simplest form of the binding. We'd prefer introduce these
properties for someone with a different use case so they can test
on their hardware.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>