CHROMIUM: img-rogue: 1.13: drop dev_priv_size from pvr_drm
With v5.11, the "dev_priv_size" variable is only available for DRM Legacy
drivers (CONFIG_DRM_LEGACY). This causes build errors as the PVR GPU driver
explicitly initializes this member to 0 (as part of initializing drm_driver
struct).

This CL removes the initialization of this member altogether (instead of
making it conditionally compiled for older versions). This is because this
member was being initialized to 0, and we don't need explicit initialization
for that.

BUG=b:190342046
TEST=build 5.11 kernel with img-rogue 1.13

Change-Id: I82ba6d7ddcc41023ebe4b83874e358daee616e41
Signed-off-by: Luigi Santivetti <luigi.santivetti@imagination.corp-partner.google.com>