HACK: pci: pcie-mediatek-gen3: Use longer timeout, retry probe if link down
Sometimes the PCIe port seems to be stuck and requiring more time to
get up... and if it still times out (which is a rare event), probing
it again seems to make it work.

As a way to get it up right now, just apply both workarounds.
In any case, a better research must be done here, as this kind of WA
is not upstreamable.