drm: mediatek: mtk_dp: Always initialize mutex update_plugged_status_lock
Function mtk_dp_update_plugged_status() uses this mutex and gets
called by mtk_dp_bridge_atomic_enable() which, in turn, is assigned
as (*atomic_enable)() callback in mtk_dp_bridge_funcs: this is
registered for both DP and eDP cases and makes use of such mutex
in both situations.

For this reason, it is necessary to move the initialization of the
mutex update_plugged_status_lock out of the !mtk_dp_is_edp() guard.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>