drm/panfrost: Add GPU IDs for Mali-G57 on MT8192 and MT8195
MediaTek MT8192 and MT8195 have a Mali-G57 with a special GPU ID.
Add these GPU IDs, but treat them as otherwise identical to a
standard Mali-G57.

We do _not_ fix up the GPU ID here -- userspace needs to be aware of the
special GPU ID, in case we find functional differences between
MediaTek's implementation and the standard Mali-G57 down the line.

[Angelo: Added GPU ID for MT8195's Mali G57]
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>