media: uapi: Fix AV1 reference lists
We need both all the references and the per frame references. This is
because unlike VP9, show_existing_frame that are key frame will refresh
the frame slots, which prevents the driver from keeping track since
these frame are not signalled to the driver.

Instead mimic other API (like VA and DXVA) and provide a complete list
of active references. The driver can then cleanup any data it was
storing for references that are no longer in use, saving previous
memory.