ASoC: dt-bindings: mediatek: mt8195: Specify phandles to codecs
Replace "mediatek,dptx-codec" and "mediatek,hdmi-codec" by introducing
a playback-codecs object, where it is possible to specify both of those
and the speaker amplifier codec(s).
While at it, also add a headset-codec object for headset codec(s) to
follow the same bindings as the newly introduced mt8186 platforms.

This commit breaks the ABI, but this is possible because currently
there is no devicetree using this binding.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>