cpufreq: mediatek: Fix KP and lockups on proc/sram regulators error
Function regulator_get_optional() returns a negative error number on
any kind of regulator_get() failure: failing to check for that in the
teardown path will lead to a kernel panic due to a call to function
regulator_disable().

Besides that, the "proc" regulator does actually provide power to the
CPU cluster(s): disabling it will produce a lockup on at least some
SoCs, such as MT8173.

That consideration is also valid for the "sram" regulator, providing
power to the CPU caches instead, present on some other SoCs, such as
MT8183, MT8186 (and others).

Resolve both situations and by simply removing the entire faulty
branches responsible for disabling the aforementioned regulators if
enabled, keeping in mind that these are enabled (and left enabled)
by the bootloader before booting the kernel.

Fixes: 4b9ceb757bbb ("cpufreq: mediatek: Enable clocks and regulators")
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>