drm/bridge: anx7625: Notify audio framework of connector status changes
We already have anx7625_audio_update_connector_status() to notify the
audio framework when HDMI is (dis)connected, but up until now this
function was only called during the callback setup. Add a call for it
to the HPD change handling logic.

A couple functions needed to be hoisted as part of this.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>

Series-to: Robert Foss <robert.foss@linaro.org>
Series-cc: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Series-cc: kernel@collabora.com